Remove defaultConfig altogether from build.gradle
Note: this change actually affects the project quite a bit. A missing
minSdkVersion falls back to the default of "1", meaning the app is
compatible with all versions of Android.

Visually this means that the most basic theme is applied to the app
(i.e. bye bye Holo).

But hey, we're going minimal after all.